requirements.txt
PyCryptodome
pytest
xades